neuroquantum_cluster::cluster_manager

Struct ClusterManager

pub struct ClusterManager { /* private fields */ }

High-level cluster manager that coordinates all components.

Implementations

impl ClusterManager

pub async fn new(config: ClusterConfig) -> ClusterResult<Self>

Create a new cluster manager.

Errors

Returns an error if the cluster node cannot be initialized.

pub async fn start(&self) -> ClusterResult<()>

Start the cluster manager.

This will:

Start the underlying cluster node

Begin background health monitoring

Start metrics collection

pub async fn stop(&self) -> ClusterResult<()>

Stop the cluster manager gracefully.

pub async fn status(&self) -> ClusterStatus

Get the current cluster status.

pub fn node(&self) -> Arc<ClusterNode>

Get the underlying cluster node.

pub fn shard_manager(&self) -> Arc<ShardManager>

Get the shard manager.

pub fn replication_manager(&self) -> Arc<ReplicationManager>

Get the replication manager.

pub fn metrics(&self) -> Arc<ClusterMetrics>

Get the cluster metrics.

pub fn consensus(&self) -> Arc<RaftConsensus>

Get the consensus module.

pub async fn is_leader(&self) -> bool

Check if this node is the leader.

pub fn node_id(&self) -> NodeId

Get the node ID.

pub async fn propose(&self, command: Vec<u8>) -> ClusterResult<u64>

Propose a command to be replicated across the cluster.

Errors

Returns an error if:

This node is not the leader

The command cannot be proposed