neuroquantum_cluster::network::proto::cluster_node_client

Struct ClusterNodeClient

pub struct ClusterNodeClient<T> { /* private fields */ }

Cluster node service for inter-node communication

Implementations

impl ClusterNodeClient<Channel>

pub async fn connect<D>(dst: D) -> Result<Self, Error>
where D: TryInto<Endpoint>, D::Error: Into<StdError>,

Attempt to create a new client by connecting to a given endpoint.

impl<T> ClusterNodeClient<T>
where T: GrpcService<Body>, T::Error: Into<StdError>, T::ResponseBody: Body<Data = Bytes> + Send + 'static, <T::ResponseBody as Body>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_origin(inner: T, origin: Uri) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F, ) -> ClusterNodeClient<InterceptedService<T, F>>
where F: Interceptor, T::ResponseBody: Default, T: Service<Request<Body>, Response = Response<<T as GrpcService<Body>>::ResponseBody>>, <T as Service<Request<Body>>>::Error: Into<StdError> + Send + Sync,

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress requests with the given encoding.

This requires the server to support it otherwise it might respond with an error.

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing responses.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX

pub async fn handshake( &mut self, request: impl IntoRequest<HandshakeRequest>, ) -> Result<Response<HandshakeResponse>, Status>

Handshake to establish connection and exchange node IDs

pub async fn append_entries( &mut self, request: impl IntoRequest<AppendEntriesRequest>, ) -> Result<Response<AppendEntriesResponse>, Status>

Append entries (log replication or heartbeat)

pub async fn request_vote( &mut self, request: impl IntoRequest<VoteRequest>, ) -> Result<Response<VoteResponse>, Status>

Request vote during leader election

pub async fn heartbeat( &mut self, request: impl IntoRequest<HeartbeatRequest>, ) -> Result<Response<HeartbeatResponse>, Status>

Heartbeat for health check

pub async fn install_snapshot( &mut self, request: impl IntoRequest<InstallSnapshotRequest>, ) -> Result<Response<InstallSnapshotResponse>, Status>

Install snapshot for new or lagging nodes